customer-database-simulator/requirements.txt
Flask==3.1.2
pyzmq==27.1.0
Faker==37.8.0
gunicorn==23.0.0